Create release.yaml(auto release notes)
added a few categories to a changelog file that should update so long as the correct labels are applied to the pull request/commit/branch merge. Basically anytime the project is updated we can use this changelog file. The labels are:
- enhancement (new feature)
- fixed (bug fix)
- breaking-change (you made some kind of change that broke something)
- * (anything that's not covered by the first three)
This file can be updated later and changed as we need it though so it's not set in stone or anything.